+++ WorkoutMirrorLive/FeaturesNew/StatsTab/Child/PersonSettings/Feature/PersonSettingsFeature+Destination.swift
@@ -1,0 +1,19 @@
+//
+//  PersonSettingsFeature+Destination.swift
+//  MyFitnessJournal
+//
+//  Created by Sebastian Sciuba on 23/09/2025.
+//
+
+import ComposableArchitecture
+
+/// Implementation of `PersonSettingsFeature` destination
+extension PersonSettingsFeature {
+    
+    @Reducer
+    enum Destination {
+
+    }
+}
+
+
