+++ WorkoutMirrorLive/FeaturesNew/StatsTab/Child/PersonSettings/Feature/PersonSettingsFeature+State.swift
@@ -1,0 +1,34 @@
+//
+//  PersonSettingsFeature+State.swift
+//  MyFitnessJournal
+//
+//  Created by Sebastian Sciuba on 23/09/2025.
+//
+
+import ComposableArchitecture
+
+/// Implementation of `PersonSettingsFeature` state
+extension PersonSettingsFeature {
+    
+    @ObservableState
+    struct State {
+        
+        // MARK: - Properties
+        
+        var age: String = "-"
+        
+        var sex: String = "-"
+        
+        var height: String = "-"
+        
+        var weight: String = "-"
+        
+        var restingHeartRate: String = "-"
+        
+        // MARK: - Destination
+        
+        /// destination from PersonSettingsFeature
+        @Presents var destination: Destination.State?
+    }
+    
+}
